Updated admin layout nav links.

+++ routes/web.php
@@ -7,26 +7,20 @@ use Inertia\Inertia;
 // Controllers
 use App\Http\Controllers\ProductController;
 
-
 Route::get('/', function () {
-    return Inertia::render('Welcome', [
-        'canLogin' => Route::has('login'),
-        'canRegister' => Route::has('register'),
-        'laravelVersion' => Application::VERSION,
-        'phpVersion' => PHP_VERSION,
-    ]);
-});
+    return redirect(route('login'));
+})->name('frontpage');
 
 Route::middleware([
     'auth:sanctum',
     config('jetstream.auth_session'),
     'verified',
 ])->group(function () {
-    Route::get('/dashboard', function () {
-        return Inertia::render('Dashboard');
-    })->name('dashboard');
+    Route::get('/admin/home', function () {
+        return redirect(route('admin.products.index'));
+    })->name('admin.home');
 
-    Route::get('/products', [ProductController::class, 'index'])->name('admin.products.index');
-    Route::post('/products', [ProductController::class, 'store'])->name('admin.products.store');
-    Route::delete('/products/{product}', [ProductController::class, 'destroy'])->name('admin.products.destroy');
+    Route::get('/admin/products', [ProductController::class, 'index'])->name('admin.products.index');
+    Route::post('/admin/products', [ProductController::class, 'store'])->name('admin.products.store');
+    Route::delete('/admin/products/{product}', [ProductController::class, 'destroy'])->name('admin.products.destroy');
 });
